[#48] Feat: 일정 추가 기능 구현
