+++ src/components/Schedule/Schedule.jsx
@@ -1,21 +1,20 @@
-import { useState, useEffect, useContext } from 'react'
+import { useState, useEffect } from 'react'
 import styled from 'styled-components'
 import ScheduleItem from './ScheduleItem'
 import ShadowyBox from '@components/shared/ShadowyBox'
 import { colors } from '@styles/Colors'
 import StyledButton from '@components/shared/Button'
-import Modal, { ModalContext } from '@components/shared/Modal'
-import Horizon from '@components/shared/Horizon'
 import { fetchEmployeeTasks } from '../../mock/fetchEmployeeTasks'
-import Flex from '@components/shared/Flex'
+import Modal from '@components/shared/Modal'
+import ModalContents from './ModalContents'
 
 const getDivisionColor = (division) => {
   switch (division) {
-    case 'meeting':
+    case 'Meeting':
       return 'rgba(255, 59, 59, 0.5)'
-    case 'prepare':
+    case 'Prepare':
       return 'rgba(255, 150, 27, 0.5)'
-    case 'external':
+    case 'External':
       return 'rgba(0, 133, 255, 0.5)'
     default:
       return 'rgb(232, 232, 232)'
@@ -46,6 +45,10 @@ export default function Schedule() {
     loadTasks()
   }, [])
 
+  const handleTaskAdded = (newTask) => {
+    setScheduleItems((prev) => [...prev, newTask])
+  }
+
   if (loading) {
     return <div>Loading...</div>
   }
@@ -60,7 +63,7 @@ export default function Schedule() {
       </ScheduleList>
       <ButtonContainer>
         <Modal Trigger={Trigger}>
-          <ModalContents />
+          <ModalContents employeeId="Zrghj2Jf3CVwQ7jSOmjCXYBBlek1" onTaskAdded={handleTaskAdded} />
         </Modal>
       </ButtonContainer>
     </ScheduleContainer>
@@ -110,237 +113,8 @@ function Trigger({ onClick }) {
       text="일정 추가"
       $backgroundColor={colors.primary_200}
       $hoverColor={colors.primary_300}
-      margin="30px 0 0 0"
+      $margin="30px 0 0 0"
       onClick={onClick}
     />
-  )
-}
-
-const Button = styled.button`
-  width: 94px;
-  height: 32px;
-  box-sizing: border-box;
-  margin: 20px;
-  margin-left: 40px;
-  padding: 6px 8px;
-  background-color: #4fd1c5;
-  color: white;
-  border: none;
-  border-radius: 8px;
-
-  &:hover {
-    background-color: #00bcab;
-  }
-`
-
-function ModalContents() {
-  const [selectedColor, setSelectedColor] = useState('')
-  return (
-    <Form>
-      <Flex $justify="space-between">
-        <ModalTitle>일정 추가</ModalTitle>
-        <Category setSelectedColor={setSelectedColor} />
-      </Flex>
-      <Horizon $width="558px" $ml="20px" />
-      <Contents selectedColor={selectedColor}></Contents>
-    </Form>
-  )
-}
-
-const Form = styled.div`
-  background-color: white;
-  width: 800px;
-  height: 600px;
-  border-radius: 15px;
-`
-
-const ModalTitle = styled.h2`
-  display: block;
-  margin-block-start: 0.83em;
-  margin-block-end: 0.83em;
-  margin-inline-start: 0px;
-  margin-inline-end: 0px;
-  unicode-bidi: isolate;
-  font-size: 18px;
-  font-weight: bold;
-  padding-left: 41px;
-`
-
-function Category({ setSelectedColor }) {
-  const Box = styled(Flex)`
-    border: 1px solid #eceff5;
-    width: 40%;
-    border-radius: 5px;
-    padding: 5px 8px;
-    margin-top: 8px;
-    margin-right: 41px;
-  `
-
-  const categories = [
-    { name: 'meeting', color: 'rgba(255, 59, 59, 0.5)' },
-    { name: 'report', color: 'rgb(198, 198, 198)' },
-    { name: 'prepare', color: 'rgba(255, 150, 27, 0.5)' },
-    { name: 'external', color: 'rgba(0, 133, 255, 0.5)' },
-  ]
-
-  return (
-    <Box $justify="space-between">
-      {categories.map((category) => (
-        <ColorBlock
-          key={category.name}
-          $name={category.name}
-          $color={category.color}
-          onClick={() => setSelectedColor(category.color)}
-        />
-      ))}
-    </Box>
-  )
-
-  function ColorBlock({ $color, $name, onClick }) {
-    const ColorBox = styled(Flex)`
-      background: #f9f9fa;
-      border-radius: 5px;
-      padding: 5px;
-      cursor: pointer;
-      transition: 0.2s;
-      &:hover {
-        box-shadow: 2px 2px 3px rgba(0, 0, 0, 0.05);
-      }
-    `
-
-    const ColorPin = styled.div`
-      width: 15px;
-      height: 15px;
-      border-radius: 50%;
-      background: ${(props) => props.$color};
-      margin-bottom: 8px;
-    `
-
-    const ColorText = styled.div`
-      color: ${(props) => props.$color};
-    `
-
-    return (
-      <ColorBox $direction="column" onClick={onClick}>
-        <ColorPin $color={$color}></ColorPin>
-        <ColorText $color={$color}>{$name}</ColorText>
-      </ColorBox>
-    )
-  }
-}
-
-function Contents({ selectedColor }) {
-  const { setIsOpen } = useContext(ModalContext)
-
-  const Container = styled.div`
-    padding: 28px 40px;
-  `
-  const Title = styled.div`
-    width: 100%;
-    height: 55px;
-    box-sizing: border-box;
-    background-color: #effaf8;
-    color: black;
-    font-weight: bold;
-    line-height: 55px;
-    padding-left: 20px;
-    position: relative;
-  `
-
-  const Content = styled.div`
-    margin: 15px 0 25px;
-  `
-
-  const Select = styled.select`
-    width: 160px;
-    height: 40px;
-    box-sizing: border-box;
-    padding: 12px;
-    padding-left: 10px;
-    border: 1px solid #eceff5;
-    outline: none;
-
-    &span {
-      width: 10px;
-      height: 10px;
-    }
-  `
-
-  function DateSelect({ data }) {
-    return (
-      <Select>
-        {data.map((x) => (
-          <option key={x}>{x}</option>
-        ))}
-      </Select>
-    )
-  }
-
-  const years = Array.from({ length: 10 }, (_, i) => 2024 + i)
-  const months = Array.from({ length: 12 }, (_, i) => 1 + i)
-  const days = Array.from({ length: 31 }, (_, i) => 1 + i)
-
-  const Input = styled.input`
-    width: 100%;
-    height: 40px;
-    resize: none;
-    border: 1px solid #eceff5;
-    outline: none;
-    margin: 15px 0 25px;
-    padding: 0 5px;
-    font-size: 18px;
-  `
-
-  const CategoryCircle = styled.span`
-    position: absolute;
-    top: 50%;
-    right: 20px;
-    transform: translateY(-50%);
-    width: 20px;
-    height: 20px;
-    border-radius: 50%;
-    background: ${selectedColor};
-  `
-
-  function SubmitButton({ children }) {
-    const handleClick = () => {
-      setIsOpen(false)
-    }
-
-    return <Button onClick={handleClick}>{children}</Button>
-  }
-
-  return (
-    <Container>
-      <Title>날짜</Title>
-      <Content>
-        <Flex $gap="60px" $justify="flex-start">
-          <DateSelect data={years} />
-          <DateSelect data={months} />
-          <DateSelect data={days} />
-        </Flex>
-      </Content>
-      <Title>
-        할 일<CategoryCircle></CategoryCircle>
-      </Title>
-      <div>
-        <Input />
-        {/* <InputCircle></InputCircle> */}
-      </div>
-      <Flex $justify="space-around">
-        <Flex $direction="column" $gap="15px">
-          <Title>상태</Title>
-          <Select></Select>
-        </Flex>
-        <Flex $direction="column" $gap="15px">
-          <Title>상태</Title>
-          <Select></Select>
-        </Flex>
-      </Flex>
-      <Flex>
-        <SubmitButton>취소</SubmitButton>
-        <SubmitButton>완료</SubmitButton>
-      </Flex>
-    </Container>
   )
 }
