+++ src/mock/fetchEmployeeTasks.jsx
@@ -1,4 +1,4 @@
-import { collection, getDocs, addDoc, Timestamp } from 'firebase/firestore'
+import { collection, getDocs } from 'firebase/firestore'
 import { db } from '../firebase/firebaseConfig'
 
 export async function fetchEmployeeTasks(employeeId) {
@@ -12,22 +12,4 @@ export async function fetchEmployeeTasks(employeeId) {
   }))
 
   return tasks
-}
-
-export async function addTaskToEmployee(employeeId, taskData) {
-  const tasksRef = collection(db, 'EMPLOYEES', employeeId, 'TASKS')
-
-  const newTask = {
-    ...taskData,
-    time: Timestamp.fromDate(new Date(taskData.time)),
-  }
-
-  try {
-    const docRef = await addDoc(tasksRef, newTask)
-    console.log('Document written with ID: ', docRef.id)
-    return docRef.id
-  } catch (e) {
-    console.error('Error adding document: ', e)
-    throw e
-  }
 }
